pyproject.toml
[project]
name = "micromorph"
version = "0.1.0"
description = ""
authors = [
    {name = "Nayjest",email = "mail@vitaliy.in"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ai-microcore>=5,<6",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = [
    "tests",
]